Allow constant-foldable expressions to be used as LIKE patterns.
SD-1523 #done

This moves the conversion from LIKE => regex into the simplifier (where
it should be) and out of the compiler. Also, non-foldable expressions
are technically allowed now, but aren’t yet supported by the MongoDB
backend (we could implement a LIKE => regex converter in mapReduce to
add support).

There was also a small secondary bug in our handling of the generic `||`
operator, where constant strings would fail to typecheck. That is also fixed.
